requirements.txt
aiogram==3.15.0
aiohttp==3.11.11
anthropic==0.42.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
